+++ app/controllers/api/experiences_controller.rb
@@ -38,14 +38,21 @@ class Api::ExperiencesController < ApplicationController
     @experience.company_name = params[:company_name] || @experience.company_name
     @experience.details = params[:details] || @experience.details
 
-    @experience.save
-    render 'show.json.jbuilder'
+    if @experience.student_id == current_student.id
+      @experience.save
+      render 'show.json.jbuilder'
+    else  
+      render json: {errors: @experience.errors.full_messages}, status: :unprocessable_entity
+    end
   end
 
   def destroy
     experience = Experience.find(params[:id])
-    experience.destroy
 
-    redner json: { message: "Successfully removed experience." }
-  end
+    if @experience.student_id == current_student.id
+      experience.destroy
+      render json: { message: "Successfully removed experience." }
+    else  
+      render json: {errors: @experience.errors.full_messages}, status: :unprocessable_entity
+    end
 end
